+++ ReleaseNotes.md
@@ -1,3 +1,4 @@
 
-* 0.1.0.2   - 07.04.2021 Added PjsipExt.
-* 0.1.0.1   - 02.04.2021 Starting the project.
+* 0.2.0   - 12.04.2021 Build Nuget Package.
+* 0.1.2   - 07.04.2021 Added PjsipExt.
+* 0.1.1   - 02.04.2021 Starting the project.